requirements.txt
django==3.1.1
mysqlclient==2.0.3
djangorestframework==3.12.2
markdown==3.3.3
django-filter==2.4.0